requirements.txt
requests
python-dotenv
paho-mqtt
backports.zoneinfo; python_version < "3.9"